fix: Handler chains logging and settings
